+++ modules/core/dependency/python-ihm/ChangeLog.rst
@@ -1,5 +1,14 @@
-HEAD
-====
+0.16 - 2020-05-29
+=================
+  - :func:`ihm.reader.read` no longer discards models read from non-IHM mmCIF
+    files; they are instead placed in their own :class:`ihm.model.ModelGroup`.
+  - Bugfix: both the pure Python and C-accelerated mmCIF readers are now more
+    robust, able to handle files in binary mode (e.g. from opening a URL)
+    and in Unicode (mmCIF files are supposed to be ASCII but python-ihm should
+    handle any encoding Python supports).
+
+0.15 - 2020-04-14
+=================
   - :class:`ihm.dataset.Dataset` objects that derive from another dataset
     can now record any transformation involved; see
     :class:`ihm.dataset.TransformedDataset`.